src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/adopt-node-with-shadow-root-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: foreignDocument.getElementById(...).createShadowRoot is not a function
PASS successfullyParsed is true

TEST COMPLETE
